refactor: move dto.ts to /api and rename to api.model.ts
