+++ src/app.controller.ts
@@ -1,7 +1,7 @@
 import { Controller, Get } from '@nestjs/common';
 import { PairsService } from './api/pairs/pairs.service';
 import { TokensService } from './api/tokens/tokens.service';
-import * as dto from './dto';
+import * as dto from './api/api.model';
 import { ApiOperation, ApiResponse } from '@nestjs/swagger';
 
 @Controller('')
