+++ src/database/pair-liquidity-info-history/pair-liquidity-info-history-db.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { PrismaService } from '../prisma.service';
 import { PairLiquidityInfoHistory } from '@prisma/client';
-import { OrderQueryEnum } from '../../dto';
+import { OrderQueryEnum } from '../../api/api.model';
 
 import { ContractAddress } from '../../clients/sdk-client.model';
